refact/front/CardTrip.jsx: moves component to src/components/cardTrip/CardTrip.jsx
